add distance-to-land subroutine
moving window search that calculates distance-to-land from current pixel.
This search moves concentrically out from the center of a window that has a maximum size set by the 'ws' variable. Any pixels that have not found land within the window size have a distance-to-land value set equal to the maximum search distance plus one meter. So if the search window size is 10, and the resolution of the DEM is 30, then if no land is found within 300 meters of a given pixel, the distance to land value for that pixel will equal 301.